[geodata] Provide a new GeoData class to serve as a base class for manipulating geo-referenced data.
The core is a structured numpy array possibly containing multiple layers.
Meta data is used to give the x and y offset (Lambert 93) as well as cell size.

Methods are provided to combine several GeoData structures: concatenate on X/Y axis or add a new layer.
A GeoData class can be exported to GTiff.