[geodata] Largely revamped clustering.
A value space is clustered by splitting it into regular sized intervals of
a predefined size (e.g. 1km/h for wind velocity).
While this method could be easily improved (e.g. by smartly placing the center
of intervals), it provides a very simple method with the advantages of:
 - not being dependent on the number of values (k-means tends to split large clusters
   to minimize the summed distance)
 - not being dependent on locality
 - being exteremely fast

In addition, a method is provided to combine N clustering of different parameters
into a single clustering.
The environment (in firemodel) has a property for creating the clustering using default
values.