Separate Dubins3dPath and Dubins3dPathLength
Recover this separation. Dubins3dPathLength only calculates the expected
length of the 3d path (which is quite fast to compute). Then, the rest
of the geometrical computation is in the inherited class Dubins3dPath,
where the geometrical math is done.

There should not be more problems with the "increase cost" assertions
failing because Dubins3dPathLength returns exact path length.